+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * New function for strand shift cross-correlation plot for phantompeakqualtools.
 * New submodule to support PCA plot from the plotPCA function for deepTools (written by [@chuan-wang](https://github.com/chuan-wang/))
 * New submodule to support correlation heatmap from the plotCorrelation function for deepTools (written by [@chuan-wang](https://github.com/chuan-wang/))
+* New submodule to support output from the `bamPEFragmentSize --outRawFragmentLengths` function for deepTools (written by [@chuan-wang](https://github.com/chuan-wang/))
 
 #### New Modules:
 * [**phantompeakqualtools**](https://www.encodeproject.org/software/phantompeakqualtools)